+++ DatatronTestImageViewer1/DatatronTestImageViewer1.csproj
@@ -56,9 +56,6 @@
     <Compile Include="DataGridViewer.Designer.cs">
       <DependentUpon>DataGridViewer.cs</DependentUpon>
     </Compile>
-    <Compile Include="DatatronImageViewer.cs">
-      <SubType>Form</SubType>
-    </Compile>
     <Compile Include="DTImageViewer.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -71,25 +68,22 @@
     <Compile Include="FloatingDialogue.Designer.cs">
       <DependentUpon>FloatingDialogue.cs</DependentUpon>
     </Compile>
-    <Compile Include="Form1.cs">
+    <Compile Include="ImageViewer.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Form1.Designer.cs">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <Compile Include="ImageViewer.Designer.cs">
+      <DependentUpon>ImageViewer.cs</DependentUpon>
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="DataGridViewer.resx">
       <DependentUpon>DataGridViewer.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="DatatronImageViewer.resx">
-      <DependentUpon>DatatronImageViewer.cs</DependentUpon>
-    </EmbeddedResource>
     <EmbeddedResource Include="DTImageViewer.resx">
       <DependentUpon>DTImageViewer.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Form1.resx">
-      <DependentUpon>Form1.cs</DependentUpon>
+    <EmbeddedResource Include="ImageViewer.resx">
+      <DependentUpon>ImageViewer.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
